Fixed Issue 29, Fixed input error with parse port and cleaned up some stuff

+++ include/scan_methods/multithread_scanning.h
@@ -18,4 +18,4 @@
 
 int multithread_scanning(scan_func_t *func_info, scan_arg_t *func_arg,
                          scan_result_t *func_result, unsigned short *ports,
-                         unsigned short port_len, unsigned int workers);
+                         unsigned int port_len, unsigned int workers);
